1) Added ex 2) Changed to bigint

+++ src/main/rersources/init_db.sql
@@ -1,7 +1,7 @@
 CREATE DATABASE `manufacturers` /*!40100 DEFAULT CHARACTER SET utf8mb4 COLLATE utf8mb4_0900_ai_ci */ /*!80016 DEFAULT ENCRYPTION='N' */;
 
 CREATE TABLE `manufacturers` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
+  `id` bigint(11) NOT NULL AUTO_INCREMENT,
   `name` varchar(255) DEFAULT NULL,
   `country` varchar(255) DEFAULT NULL,
   `is_deleted` tinyint(4) NOT NULL DEFAULT '0',
